Add support for keystone v3
This will allow variables to passtrough the plugin and be relayed to
fog. Fog 1.0.3 doesn't seems to work well with keystonev2 so we dont
bump for now, further migration will be needed to support this version.

The plugin will still support a vagrant file that was working with v2
for now.