Add support for availability zone
This is needed to choose which compute node you will boot on.
Not added test to create_server, this is tech debt
we should address soon.

Uses the fog default if the availability zone is not passed, should
not break anything.

Tested locally with and without the availability_zone config.